requirements.txt
pytz
requests
python-dotenv
git+https://github.com/luxkatana/somtodayapi_python@dev